+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [7.4.0-beta.3](https://github.com/Lint-Free-Technology/uix/compare/v7.4.0-beta.2...v7.4.0-beta.3) (2026-05-28)
+
+### 🐞 Bug Fixes
+
+* Refresh forge templates on has/panel updates. Correct related section mold visibility issues ([#348](https://github.com/Lint-Free-Technology/uix/issues/348)) ([3c9bed6](https://github.com/Lint-Free-Technology/uix/commit/3c9bed65d2d91371cfcea75a8c21bf7a72f59430)), closes [#347](https://github.com/Lint-Free-Technology/uix/issues/347)
+* Update hash/panel on all history state changes including pushState and replaceState ([97cd49e](https://github.com/Lint-Free-Technology/uix/commit/97cd49ee48af83119042e7c8acba089956d3bc04)), closes [#349](https://github.com/Lint-Free-Technology/uix/issues/349)
+
+## [7.4.0-beta.2](https://github.com/Lint-Free-Technology/uix/compare/v7.4.0-beta.1...v7.4.0-beta.2) (2026-05-25)
+
+### ⭐ New Features
+
+* Allow to style states-history-charts by theme variable. ([7594b21](https://github.com/Lint-Free-Technology/uix/commit/7594b2173c1cf91c38ae53442339d5b1bff9e2e0)), closes [#342](https://github.com/Lint-Free-Technology/uix/issues/342)
+
+## [7.4.0-beta.1](https://github.com/Lint-Free-Technology/uix/compare/v7.3.0...v7.4.0-beta.1) (2026-05-19)
+
+### ⭐ New Features
+
+* Add a UIX Forge `theme` spark for local frontend theme application ([#338](https://github.com/Lint-Free-Technology/uix/issues/338)) ([dceaeb6](https://github.com/Lint-Free-Technology/uix/commit/dceaeb62d574b885ce21608a0ce7990b056bf994))
+* **styling:** Apply theme override to element with `uix.theme` ([#337](https://github.com/Lint-Free-Technology/uix/issues/337)) ([ab02453](https://github.com/Lint-Free-Technology/uix/commit/ab02453a688343ed47007bbda08675eede905b94))
+
 ## [7.3.0](https://github.com/Lint-Free-Technology/uix/compare/v7.2.0...v7.3.0) (2026-05-17)
 
 ### ⭐ New Features